+++ ofxclient/__init__.py
@@ -1,4 +1,8 @@
-from institution import Institution
-from account import Account, BrokerageAccount, CreditCardAccount, BankAccount
-from client import Client
-from version import __version__
+from __future__ import absolute_import
+from __future__ import unicode_literals
+from ofxclient.account import (
+    Account, BrokerageAccount, CreditCardAccount, BankAccount
+)
+from ofxclient.client import Client
+from ofxclient.institution import Institution
+from ofxclient.version import __version__