+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Added
+- Use unicode instead of byte string wherever possible
+- Dropping Python 2.5 support
+- Adding Python 3 support
 ### Fixed
 - Merging multiple OFX documents into one now closes tag properly.
 - LICENSE file included in dist